% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_and_predict_model.R
\name{build_and_predict_model}
\alias{build_and_predict_model}
\title{Build a Linear Model and Predict Housing Prices}
\usage{
build_and_predict_model(data, formula)
}
\arguments{
\item{data}{A dataframe containing the housing data.}

\item{formula}{A formula specifying the model to be fitted.}
}
\value{
A list containing the linear model and predictions.
}
\description{
This function builds a linear regression model using the specified formula and dataset.
It returns the model and predictions.
}
