% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_price_comparison.R
\name{visualize_price_comparison}
\alias{visualize_price_comparison}
\title{Visualize Actual vs Predicted Prices}
\usage{
visualize_price_comparison(data, actual_col, predicted_col, plot_path)
}
\arguments{
\item{data}{Data frame containing actual and predicted prices.}

\item{actual_col}{Name of the column containing actual prices.}

\item{predicted_col}{Name of the column containing predicted prices.}

\item{plot_path}{Path where the plot image will be saved.}
}
\description{
Generates a scatter plot comparing actual and predicted housing prices.
}
